RT4408: Remove last traces of CRYPTO_dynlock for non-compatibility build
Commit 2e52e7df5 ("Remove the old threading API") left a dummy definition
of the CRYPTO_dynlock for compatibility, if OPENSSL_API_COMPAT < 1.1.0.

However, there's still a DEFINE_STACK_OF(CRYPTO_dynlock) in cryptlib.h
which isn't so masked, and breaks the build if you disable the API
compatibility. Assuming that's supposed to be present too for the same
reason, wrap it in a similar #ifdef.