add key type rsaesOaep
All TPM modules come with an x509 certificate in their NV ram.  This is
used to verify the build in endorsement key was genuinely provisioned
by the manufacturer and verifying this certificate is a key step in the
chain to authenticating you're talking to a genuine TPM and not a fake
one.  Unfortunately, because of TPM specs, the signatures the TPM does
in the quotes are all of type rsaeaOaep and openssl currently gives
this error whenever it sees this key type (using openssl x509 -text -in
tpm.crt):

Certificate:
    Data:
        Version: 3 (0x2)
        Serial Number: 4432329820887552212 (0x3d82cc7d155f90d4)
    Signature Algorithm: sha1WithRSAEncryption
        Issuer: CN=Nuvoton TPM Root CA 2010, O=Nuvoton Technology Corporation, C=TW
        Validity
            Not Before: Jun 23 16:58:33 2015 GMT
            Not After : Jun 19 16:58:33 2035 GMT
        Subject:
        Subject Public Key Info:
            Public Key Algorithm: rsaesOaep
            Unable to load Public Key
140164248934032:error:0609E09C:digital envelope routines:PKEY_SET_TYPE:unsupported algorithm:p_lib.c:239:
140164248934032:error:0B07706F:x509 certificate routines:X509_PUBKEY_get:unsupported algorithm:x_pubkey.c:155:

openssl actually understands the OAEP padding, it's just that the key
type is not understood, so this simple patch adds the key type as an
alias of the standard RSA key type.

Signed-off-by: James Bottomley <jejb@linux.vnet.ibm.com>